# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation for the Bookstore API",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/karyawan": {
            "get": {
                "tags": [
                    "Karyawan"
                ],
                "summary": "Ambil data karyawan, dengan opsi pencarian berdasarkan nama",
                "operationId": "GetKaryawan",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Kata kunci pencarian nama karyawan (opsional)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data karyawan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data karyawan tersedia"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nama": {
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    },
                                                    "nip": {
                                                        "type": "string",
                                                        "example": "6535534"
                                                    },
                                                    "jenis_kelamin": {
                                                        "type": "string",
                                                        "example": "pria"
                                                    },
                                                    "tanggal_lahir": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2025-06-01"
                                                    },
                                                    "alamat": {
                                                        "type": "string",
                                                        "example": "Jl. Contoh No. 123"
                                                    },
                                                    "jabatan": {
                                                        "type": "string",
                                                        "example": "manager"
                                                    },
                                                    "tunjangan": {
                                                        "type": "number",
                                                        "example": 1500000
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-06-02T10:00:00Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-06-02T11:00:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/karyawan/create": {
            "post": {
                "tags": [
                    "Karyawan"
                ],
                "summary": "Tambah data karyawan baru",
                "operationId": "CreateKaryawan",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama",
                                    "nip",
                                    "jenis_kelamin",
                                    "tanggal_lahir",
                                    "jabatan",
                                    "tunjangan"
                                ],
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "Budi Santoso"
                                    },
                                    "nip": {
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "jenis_kelamin": {
                                        "type": "string",
                                        "enum": [
                                            "pria",
                                            "perempuan"
                                        ],
                                        "example": "pria"
                                    },
                                    "tanggal_lahir": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-05-01"
                                    },
                                    "alamat": {
                                        "type": "string",
                                        "example": "Jl. Melati No. 12"
                                    },
                                    "jabatan": {
                                        "type": "string",
                                        "enum": [
                                            "manager",
                                            "kepala divisi",
                                            "karyawan tetap",
                                            "karyawan kontrak",
                                            "office boy"
                                        ],
                                        "example": "manager"
                                    },
                                    "tunjangan": {
                                        "type": "integer",
                                        "example": 2000000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Karyawan berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Karyawan berhasil ditambahkan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validasi gagal"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/karyawan/delete/{id}": {
            "delete": {
                "tags": [
                    "Karyawan"
                ],
                "summary": "Hapus karyawan berdasarkan ID",
                "operationId": "DeleteKaryawan",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID karyawan yang akan dihapus",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data Karyawan berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data Karyawan berhasil dihapus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Gagal menghapus data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Karyawan tidak ditemukan atau gagal menghapus data"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": {
        "security": [
            {
                "bearerAuth": []
            }
        ]
    }
}